+++ package.json
@@ -34,9 +34,11 @@
     "@nx/eslint": "17.3.0-beta.8",
     "@nx/eslint-plugin": "17.3.0-beta.8",
     "@nx/storybook": "17.3.0-beta.8",
+    "@types/conventional-recommended-bump": "^9.0.3",
     "@types/css-font-loading-module": "^0.0.7",
     "@types/fs-extra": "^9.0.13",
     "@types/jest": "26.0.24",
+    "@types/shelljs": "^0.8.15",
     "@typescript-eslint/eslint-plugin": "6.13.2",
     "@typescript-eslint/parser": "6.13.2",
     "camelcase": "^6.2.0",
@@ -47,8 +49,8 @@
     "eslint": "^8.51.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-no-null": "^1.0.2",
-    "eslint-plugin-react": "^7.32.1",
     "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-react": "^7.32.1",
     "fs-extra": "^9.0.0",
     "jest": "26.6.3",
     "jest-cli": "26.6.3",
