+++ project.json
@@ -2,18 +2,22 @@
   "name": "workspace",
   "$schema": "node_modules/nx/schemas/project-schema.json",
   "targets": {
-    "release": {
+    "publish": {
       "executor": "nx:run-commands",
       "dependsOn": ["^build"],
       "options": {
-        "command": "node ./scripts/release.js dryRun={args.dryRun} canary={args.canary}"
+        "command": "ts-node ./scripts/publish.ts dryRun={args.dryRun} releaseType={args.releaseType}"
       },
+      "defaultConfiguration": "dryRun",
       "configurations": {
-        "canary": {
-          "args": "--canary=true"
+        "release": {
+          "args": "--releaseType=release --dryRun=true"
+        },
+        "prerelease": {
+          "args": "--releaseType=prerelease --dryRun=true"
         },
         "dryRun": {
-          "args": "--dryRun=true"
+          "args": "--releaseType=release --dryRun=true"
         }
       }
     }
